// WEB/src/app/modules/giohang/model/thanhtoan-donhang-view-model.ts
export class ThanhToanDonHangViewModel {
    ID_HinhThuc_ThanhToan: number = 0;
    NguoiNhan_HoTen: string = '';
    NguoiNhan_SoDienThoai: string = '';
    TongTienDonHang: number = 0;
    IsThanhToan: boolean = false;
    SoNha: string = '';
    DiaChiNhanHang: string = '';
    ID_TinhThanh: number | undefined;
    ID_QuanHuyen: number | undefined;
    ID_PhuongXa: number | undefined;
    GhiChu: string = '';
    listID_DonHang: string = '';
}